Refactor engine: Extract symbol queries and hot reload to separate modules
Split refactor-engine.ts (2312 lines) into focused modules:
- symbol-queries.ts: Symbol location, validation, and occurrence gathering
- hot-reload.ts: Hot reload cascade, safety checks, and patch generation
- refactor-engine.ts: Core rename planning and validation (reduced to 1560 lines)

This improves code organization by applying ISP principles, making each module
responsible for a single cohesive domain. All existing tests pass without modification.
